+++ examples/rng-display.rs
@@ -105,7 +105,7 @@ fn main() -> ! {
             }
             disp.flush().unwrap();
             //delay a little while between refreshes so the display is readable
-            delay_source.delay_ms(100u8);
+            delay_source.delay_ms(100);
         }
     }