+++ src/timer/hal_1.rs
@@ -4,7 +4,7 @@
 //! a 16-bit prescaler.
 
 use core::convert::Infallible;
-use embedded_hal_one::delay::DelayNs;
+use embedded_hal::delay::DelayNs;
 
 use super::{Delay, Instance, PwmChannel, SysDelay, WithPwm};
 use fugit::ExtU32Ceil;
@@ -33,13 +33,11 @@ impl<TIM: Instance, const FREQ: u32> DelayNs for Delay<TIM, FREQ> {
     }
 }
 
-impl<TIM: Instance + WithPwm, const C: u8> embedded_hal_one::pwm::ErrorType for PwmChannel<TIM, C> {
+impl<TIM: Instance + WithPwm, const C: u8> embedded_hal::pwm::ErrorType for PwmChannel<TIM, C> {
     type Error = Infallible;
 }
 
-impl<TIM: Instance + WithPwm, const C: u8> embedded_hal_one::pwm::SetDutyCycle
-    for PwmChannel<TIM, C>
-{
+impl<TIM: Instance + WithPwm, const C: u8> embedded_hal::pwm::SetDutyCycle for PwmChannel<TIM, C> {
     fn max_duty_cycle(&self) -> u16 {
         self.get_max_duty()
     }